change grad from i32 to i64

+++ common/src/cond_stmt_base.rs
@@ -20,8 +20,8 @@ pub struct CondStmtBase {
     pub arg1: u64,
     pub arg2: u64,
 
-    pub grad1: i32,
-    pub grad2: i32,
+    pub grad1: i64,
+    pub grad2: i64,
 }
 
 /*
